fix(ci): python-coverage-shards — split oversized test_profiles.py under 1500-line maintainability limit
The git-auth work pushed tests/unit/profiles/test_profiles.py to 1631
lines, tripping test_first_party_code_files_stay_under_line_limit
(MAX_FIRST_PARTY_FILE_LINES = 1500). Extract the self-contained
merge_agent_environment git-config re-indexing tests into a new
tests/unit/profiles/test_compose_merge_environment.py, bringing
test_profiles.py to 1438 lines. No test behavior changes.
